+++ src/js/ui/NewDataExploration/BrowseDatasets.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import LayerNavigationControl from "./LayerNavigationControl";
 import {Card} from "@material-ui/core";
-import {createSection} from "./DatasetDisplayHelpers";
+import IndividualLayer from "./IndividualLayer";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -13,6 +13,18 @@ const useStyles = makeStyles((theme) => ({
         maxHeight: "65vh",
     },
 }));
+
+function createSection(datasets, workspace, setWorkspace, openLayers, setOpenLayers, booleanWorkspace, setBooleanWorkspace) {
+    let layerSection = [];
+    datasets.map((layer) =>
+        layerSection.push(
+            <IndividualLayer layer={layer} workspace={workspace} setWorkspace={setWorkspace}
+                             openLayers={openLayers} setOpenLayers={setOpenLayers}
+                             booleanWorkspace={booleanWorkspace} setBooleanWorkspace={setBooleanWorkspace} datasets={datasets}/>
+        )
+    )
+    return layerSection;
+}
 
 export default function BrowseDatasets(props) {
     const classes = useStyles();
